Adds session switching, improved interface, and a new feature that lets users set the default path of a newly created session
