+++ start_session_manager.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python2
+import tmux_session_manager
+
+tmux_session_manager.main()
